@@ -2,7 +2,7 @@ import * as React from 'react'
 import { AppState } from '../app'
 import { getActiveWallet  } from '../helpers/wallets';
 import { BalancesPresentation } from './BalancesPresentation';
-const StellarSdk = require('stellar-sdk')
+import * as StellarSdk from 'stellar-sdk'
 
 export class Balances extends React.Component<{appState: AppState}, {account: any, kinesisBalance: number, accountActivated: boolean}> {
   constructor (props) {
@@ -29,7 +29,7 @@ export class Balances extends React.Component<{appState: AppState}, {account: an
     try {
       const server = new StellarSdk.Server(props.appState.connection.horizonServer, {allowHttp: true})
       const account = await server.loadAccount(getActiveWallet(props.appState).publicKey)
-      const kinesisBalance = account.balances.filter(b => b.asset_type === 'native')[0].balance
+      const kinesisBalance = Number(account.balances.filter(b => b.asset_type === 'native')[0].balance)
       this.setState({account, kinesisBalance, accountActivated: true})
     } catch (e) {
       this.setState({accountActivated: false, kinesisBalance: 0})
@@ -38,7 +38,11 @@ export class Balances extends React.Component<{appState: AppState}, {account: an
 
   render() {
     return (
-      <BalancesPresentation appState={this.props.appState} kinesisBalance={this.state.kinesisBalance} accountActivated={this.state.accountActivated} />
+      <BalancesPresentation
+        appState={this.props.appState}
+        kinesisBalance={this.state.kinesisBalance}
+        accountActivated={this.state.accountActivated}
+      />
     )
   }
 }
