@@ -2,15 +2,17 @@ import * as React from 'react'
 import { AppState } from '../app'
 import { getActiveWallet } from '../helpers/wallets';
 import { DashboardPresentation } from './DashboardPresentation';
-const StellarSdk = require('stellar-sdk')
+import * as StellarSdk from 'stellar-sdk'
 
-export class Dashboard extends React.Component<{appState: AppState, setWalletList: Function, changeView: Function, setPassword: Function}, {account: any, transfering: boolean}> {
+export class Dashboard extends React.Component<
+  {appState: AppState, setWalletList: Function, changeView: Function, setPassword: Function},
+  {account: StellarSdk.AccountResponse | null, transfering: boolean, transferViewState: 'multi' | 'payment'}> {
   public tx
   public balances
 
   constructor (props) {
     super(props)
-    this.state = { account: null, transfering: false }
+    this.state = { account: null, transfering: false, transferViewState: 'payment'}
   }
 
   async componentDidMount() {
@@ -20,6 +22,7 @@ export class Dashboard extends React.Component<{appState: AppState, setWalletLis
   componentWillReceiveProps(nextProps) {
     if (this.props !== nextProps) {
       this.loadAccount(nextProps)
+      this.setState({transferViewState: 'payment'})
     }
   }
 
@@ -47,6 +50,11 @@ export class Dashboard extends React.Component<{appState: AppState, setWalletLis
     }
   }
 
+  public updateTransferView = (view: 'multi' | 'payment') => {
+    this.setState({transferViewState: view})
+  }
+
+
   render() {
     return (
       <DashboardPresentation
@@ -59,6 +67,8 @@ export class Dashboard extends React.Component<{appState: AppState, setWalletLis
         transfering={this.state.transfering}
         tx={(tx) => {this.tx = tx}}
         balances={(balances) => {this.balances = balances}}
+        transferView={this.state.transferViewState}
+        updateTransferView={this.updateTransferView}
       />
     )
   }
